refactor: replace getUserInfos with getUser
