+++ apps/web/app/(app)/milestones/page.tsx
@@ -1,17 +1,17 @@
-import { AppHeader } from "~/components/app-header";
+import { Layout, LayoutContent, LayoutHeader } from "~/components/layouts/layout";
 
 export default function MilestonesPage() {
 	return (
-		<>
-			<AppHeader items={["Advanced", "Milestones"]} />
-			<div className="flex flex-1 flex-col gap-4 p-4 pt-0">
+		<Layout>
+			<LayoutHeader items={["Advanced", "Milestones"]} />
+			<LayoutContent>
 				<div className="grid auto-rows-min gap-4 md:grid-cols-3">
 					<div className="aspect-video rounded-xl bg-muted/50" />
 					<div className="aspect-video rounded-xl bg-muted/50" />
 					<div className="aspect-video rounded-xl bg-muted/50" />
 				</div>
 				<div className="min-h-screen flex-1 rounded-xl bg-muted/50 md:min-h-min" />
-			</div>
-		</>
+			</LayoutContent>
+		</Layout>
 	);
 }
