+++ apps/web/app/(app)/overview/page.tsx
@@ -1,6 +1,8 @@
 
 import { Suspense } from "react";
 
+import { getUser } from "@repo/auth";
+
 import { DateRangeSelector, DateRangeSelectorSkeleton } from "~/components/date-range-selector/date-range-selector";
 import { Layout, LayoutContent, LayoutHeader } from "~/components/layouts/layout";
 import { UserHasNotPackage } from "~/components/user-has-not-package";
@@ -9,10 +11,9 @@ import { ListeningPatternChart, ListeningPatternChartSkeleton } from "~/features
 import { StatsCards, StatsCardsSkeleton } from "~/features/overview/components/stats-cards";
 import { TopArtistsCard } from "~/features/overview/components/top-artists-card";
 import { TopTracksCard } from "~/features/overview/components/top-tracks-card";
-import { getUserInfos } from "~/lib/utils-server";
 
 export default async function OverviewPage() {
-	const { userId, isDemo, hasPackage } = await getUserInfos();
+	const { userId, isDemo, hasPackage } = await getUser();
 
 	if (!hasPackage)
 		return (
