+++ apps/web/app/(app)/rankings/albums/page.tsx
@@ -1,13 +1,14 @@
 import { Suspense } from "react";
 
+import { getUser } from "@repo/auth";
+
 import { DateRangeSelector, DateRangeSelectorSkeleton } from "~/components/date-range-selector/date-range-selector";
 import { Layout, LayoutContent, LayoutHeader } from "~/components/layouts/layout";
 import { RankingAlbums } from "~/features/rankings/components/ranking-albums";
 import { SelectListLayout } from "~/features/stats/components/select-list-layout";
-import { getUserInfos } from "~/lib/utils-server";
 
 export default async function RankingsAlbumsPage() {
-	const { userId, isDemo } = await getUserInfos();
+	const { userId, isDemo } = await getUser();
 
 	return (
 		<Layout>
