+++ apps/web/app/(app)/recently-played/page.tsx
@@ -1,10 +1,11 @@
+import { getUser } from "@repo/auth";
+
 import { Layout, LayoutContent, LayoutHeader } from "~/components/layouts/layout";
 import { RecentlyPlayed } from "~/features/stats/components/recently-played";
 import { SelectListLayout } from "~/features/stats/components/select-list-layout";
-import { getUserInfos } from "~/lib/utils-server";
 
 export default async function RecentlyPlayedPage() {
-	const { userId } = await getUserInfos();
+	const { userId } = await getUser();
 
 	return (
 		<Layout>
