+++ apps/web/app/(app)/stats/activity/page.tsx
@@ -1,5 +1,7 @@
 import { Suspense } from "react";
 
+import { getUser } from "@repo/auth";
+
 import {
 	DateRangeSelector,
 	DateRangeSelectorSkeleton,
@@ -21,10 +23,9 @@ import {
 	TimeListenedChart,
 	TimeListenedChartSkeleton,
 } from "~/features/activity/components/time-listened-chart";
-import { getUserInfos } from "~/lib/utils-server";
 
 export default async function StatsActivityPage() {
-	const { userId, isDemo } = await getUserInfos();
+	const { userId, isDemo } = await getUser();
 
 	return (
 		<Layout>
