+++ apps/web/app/(app)/top/artists/page.tsx
@@ -1,14 +1,15 @@
 import { Suspense } from "react";
 
+import { getUser } from "@repo/auth";
+
 import { Layout, LayoutContent, LayoutHeader } from "~/components/layouts/layout";
 import { SelectListLayout } from "~/features/stats/components/select-list-layout";
 import { TimeRangeInfo } from "~/features/stats/components/time-range-info";
 import { TimeRangeSelect, TimeRangeSelectSkeleton } from "~/features/stats/components/time-range-select";
 import { TopArtists } from "~/features/stats/components/top-artists";
-import { getUserInfos } from "~/lib/utils-server";
 
 export default async function TopArtistsPage() {
-	const { userId, isDemo } = await getUserInfos();
+	const { userId, isDemo } = await getUser();
 
 	return (
 		<Layout>
