+++ apps/web/components/lists/music-layout/index.tsx
@@ -1,16 +1,9 @@
 "use client";
 
-import { Info } from "lucide-react";
-
-import { Alert, AlertDescription, AlertTitle } from "@repo/ui/alert";
-import { cn } from "@repo/ui/lib/utils";
-
-import { MusicItemCard } from "~/components/cards/music-item-card";
 import type { MusicItemCardProps } from "~/components/cards/music-item-card/type";
 import { useListLayout } from "~/lib/store";
 
-import type { MusicListConfig } from "../music-list/config";
-import { MusicListError } from "../music-list/error";
+import { MusicList, type MusicListConfig } from "../music-list";
 
 type MusicLayoutProps = {
     data: Array<MusicItemCardProps["item"]> | null;
@@ -19,33 +12,6 @@ type MusicLayoutProps = {
 
 export const MusicLayout = ({ data, config }: MusicLayoutProps) => {
     const listLayout = useListLayout((state) => state.list_layout);
-    const layout = config.layout ?? listLayout;
 
-    // TODO: handle error
-    if (!data) return <MusicListError />;
-
-    return (
-        <div className={cn(layout === "grid" ? "grid grid-cols-2 xs:grid-cols-3 gap-4 sm:grid-cols-4 md:grid-cols-5 lg:grid-cols-6 xl:grid-cols-7" : "flex flex-col")}>
-            {data.map((item, index) => (
-                <MusicItemCard
-                    key={`${item.id}-${index}-${config.label}`}
-                    item={item}
-                    rank={config.showRank ? index + 1 : undefined}
-                    showAction={!!config.actionHref}
-                    showHistoricalRankings={config.showHistoricalRankings}
-                    actionHref={config.actionHref}
-                    layout={layout}
-                />
-            ))}
-            {data.length === 0 && (
-                <Alert variant="info" className="col-span-full">
-                    <Info className="size-4" />
-                    <AlertTitle>No {config.label} found</AlertTitle>
-                    <AlertDescription>
-                        You haven't listened to any music during this period
-                    </AlertDescription>
-                </Alert>
-            )}
-        </div>
-    );
+    return <MusicList data={data} config={{ ...config, layout: config.layout ?? listLayout }} />
 };
