--- apps/web/components/lists/music-list/config.ts
@@ -1,9 +1,0 @@
-import type { ACTIONS_HREF } from "~/components/cards/music-item-card";
-
-export type MusicListConfig = {
-	label: string;
-	actionHref?: keyof typeof ACTIONS_HREF;
-	showRank?: boolean;
-	showHistoricalRankings?: boolean;
-	layout?: "grid" | "list";
-};
