+++ apps/web/components/lists/music-list/skeleton.tsx
@@ -3,17 +3,17 @@ import { Separator } from "@repo/ui/separator";
 
 import { CardSkeleton } from "~/components/cards/music-item-card/skeleton";
 
-type MusicLayoutSkeletonProps = {
+type MusicListSkeletonProps = {
     length?: number;
     layout?: "grid" | "list";
     showRank?: boolean;
 };
 
-export const MusicLayoutSkeleton = ({
+export const MusicListSkeleton = ({
     length = 50,
     layout = "list",
     showRank = true,
-}: MusicLayoutSkeletonProps) => {
+}: MusicListSkeletonProps) => {
     return (
         <div className={cn(layout === "grid" ? "grid grid-cols-2 xs:grid-cols-3 gap-4 sm:grid-cols-4 md:grid-cols-5 lg:grid-cols-6 xl:grid-cols-7" : "flex flex-col")}>
             {Array.from({ length }).map((_, index) => (
