+++ apps/web/features/rankings/components/ranking-albums.tsx
@@ -1,5 +1,5 @@
 import { MusicLayout } from "~/components/lists/music-layout";
-import type { MusicListConfig } from "~/components/lists/music-list/config";
+import type { MusicListConfig } from "~/components/lists/music-list";
 
 import { getRankingAlbumsData } from "../data/ranking-albums";
