+++ apps/web/features/rankings/components/ranking-artists.tsx
@@ -1,5 +1,5 @@
 import { MusicLayout } from "~/components/lists/music-layout";
-import type { MusicListConfig } from "~/components/lists/music-list/config";
+import type { MusicListConfig } from "~/components/lists/music-list";
 
 import { getRankingArtistsData } from "../data/ranking-artists";
