+++ apps/web/features/rankings/components/ranking-tracks.tsx
@@ -1,5 +1,5 @@
 import { MusicLayout } from "~/components/lists/music-layout";
-import type { MusicListConfig } from "~/components/lists/music-list/config";
+import type { MusicListConfig } from "~/components/lists/music-list";
 
 import { getRankingTracksData } from "../data/ranking-tracks";
