+++ apps/web/features/stats/components/top-artists.tsx
@@ -1,6 +1,6 @@
 import { MusicLayout } from "~/components/lists/music-layout";
-import type { MusicListConfig } from "~/components/lists/music-list/config";
-import { tryCatch } from "~/lib/utils-server";
+import type { MusicListConfig } from "~/components/lists/music-list";
+import { tryCatch } from "~/lib/utils";
 import { getHistoricalArtistRankings } from "~/services/historical-rankings";
 
 import { getTopArtists } from "../data/top-artists";
