+++ apps/web/features/stats/components/top-tracks.tsx
@@ -1,6 +1,6 @@
 import { MusicLayout } from "~/components/lists/music-layout";
-import type { MusicListConfig } from "~/components/lists/music-list/config";
-import { tryCatch } from "~/lib/utils-server";
+import type { MusicListConfig } from "~/components/lists/music-list";
+import { tryCatch } from "~/lib/utils";
 import { getHistoricalTrackRankings } from "~/services/historical-rankings";
 
 import { getTopTracks } from "../data/top-tracks";
